CUI and interface to business layer
Note to myself: boolean doneWithStep in CUI bothers me, in general I feel as if the class could be written more efficiently. The question is, how I can achieve that the player can repeat something as long as he or she wants until he or she decides to do the next step. Is there a better way than how I did it now?