Interface
I think there are too many methods in the interface and they are not all related to each other. Furthermore some seem weird to me, especially the ones I changed right now, they are all regarding the countries - should I not better just have a map with all the countries and owners in the presentation layer to keep track? Is this valid in the presentation?