Lost with country structure
I think I knew it all the time but something is wrong with the way World, IWorldManager and the management of the countries in Game (and now also Mission) is being structured. I feel like rewriting all of World and IWorldManager.

When the Army of a country changes, it needs to be displayed in the respective map where the info is saved. Is it even sensible to have a class called Continent? When I have several maps containing Country objects and furthermore in the Country class other neighbouring countries and the continent, how can I ensure that in all of these Country objects the information of the owner is being displayable? Does it even have to be displayable in every map, is it not sufficient to have this info in one map? But then, both countryMap and continents in World should contain this info, how could I otherwise check in Mission if the world has been conquered?

I think I think way too complicated, there is a structural problem in my code and it should be structured easier. Do I have to change my class attributes?