less confused - only one map tracks who conquered what
Map in Game shall track who captured which country. World has no knowledge of captured countries. Mission uses setCountries to keep track of captured countries, the derived classes can have a map for comparison (capture world = map with all countries of the world). Maybe this is an option. Could work but I am not sure if it is a good way.