Exceptions implemented and included in CUI and business logic
Not sure if I did not create unnecessary exceptions and is it clever to have one exception for a unit amount that is too high and one that is too low? Also I am really unsure about the handling of ExceptionObjectDoesntExist in receiveUnits in the CUI, will it not create an endless loop to then let the next player play? What can I do if this error occurs? And in World and PlayerManager, do I need there exceptions too and print the stacktrace? When precisely do I write an exception, it's maybe not just for informing the user how I did it now. To be honest, I dont know for how many more exceptions I have the motivation but I should at least always check somehow if a list that is queried is empty and things like that.