FilePersistence
not sure if this is the most efficient way, should I in any way split fetchContinents() in two? The method is quite long. Does every country to keep track of its neighbours even need a list of countries or would a list of country names not suffice? Also Mission definitely needs to be changed, it wont work like this, I have to understand inheritance and abstract classes better. Another question I have is, which place in the architecture the best place is to create an instance of the persistence - is it okay to create it in the presentation layer? I would prefer to create the instance in the business layer. I think this shows a flaw in my architecture, I wonder if I should have a second interface or anything which encapsulates those methods of the business layer which are not meant for the presentation layer but needed for the business classes to communicate with each other. Was it even a good idea to have several interfaces? Is it okay to write the name of the file in the persistence or should that happen in the business layer? Too many questions...