first testing - no surprise, system crashes
shifted currentplayer into playerManager and removed it from Game and CUI.
I wonder if I should have on GameManager and one RiskTurnManager or something/ if it is necessary to split this manager in two.

Persistence needs to be changed - the current GameState.txt is not sorted by players / I have to find a good way how to fill the Map<String, Map<String, Country>> countryPlayerMap after restarting - is it really necessary to iterate through all present maps/lists to fill this? To change the structure so that this is not necessary seems better.

The change from currentPlayer to playerManager showed me somehow how much effect this small shift can have, quite some methods didnt need any parameter anymore and got simplified. I wonder if I should not stop now and restructure more?