Thoughts
Just to write down my ideas for next time:

Maybe I could save the countryPlayerMap in GameStateTxt, the first number would determine the total of possessed countries by each player, then these would be read and saved in Map<String, Country>, still the method would need to read from fetchContinents() to get all info about each Country and from fetchPlayerGameState() to determine the amount of players which wrote in the txt.

I wonder if I make all of this a bit too complicated, this could be a possible solution but is it not a bit too complicated?