Save and reload works on the surface
To have the method getCountryFromContinent was born out of me not knowing how to handle differently, I think the whole persistence is a bit "patched up" and not very elegantly written right now. Furthermore I realised that I should also when saving the game save, if the user was in the step receiveunits, attack, defend, moveunits (...). This should be saved in the persistence as well but where? Seems strange to have a seperate txt, I could write it in the first line of the one I already use but is it okay to write everything in one txt? I could then keep track in a string which could be also displayed in the menu "Receive Units" in which step we were, the round number should be saved too.