keep track of round turn and step
seems a bit strange to do so in a different txt but I will leave it for now. Also I remembered, my professor said we should prevent to use integer if  I remember right and rather use enums for example because it's easier to test or something. Maybe I will have to change that, have to look this up.