change name of method
in order to make me feel as if it is a bit less game connected and has more reason to be in playermanager. I think saveGame() should from the logic be shifted to gamemanager but then, most of the infos I need to save are actually in the playermanager. Still, I would say it is the responsibility of the gamemanager, so I should shift (?)