Testing
problems: maps are not synchronized, I have to understand this better, the one in PlayerManager / the internal one in Player and the one in world are not synchronized. Furthermore, when reading from the persistence the map in world is also not automatically synchronized with the changes from setArmy in the persistence. My belief that all of them operate on the same instance of country and therefore dont need to get updated was wrong. I really need to understand this better, the maps save references to the object? If I change the state of the object with setArmy (which in itself, since there are a lot of getter before it, is probably not right, in c++ with const getters this would not be possible probably and I shouldnt do this) ?? I am quite confused, will have to read about this / inform myself.