Restructure more
I dreamed that I have to change Common completely, remove Country and use the String countryName instead. Somewhat it made me happy that all changes in common did not cause one error in the presentation / had no effect there. Restructured the persistence again, continent and country are now more seperated and continent is not that "powerful" and essential to keep track of changes in the army size but the more easy to access countryMap.

One issue remains and I dont understand why: If I conquer a country, it is not visible in the map of the player, only in the countryMap... I use changeCountryOwner when conquering. Something does not work there but at least reading and writing to and from the persistence seems to work.