Starting GUI
Color is being recognized. To read Mission from the persistence doesnt work and I dont have the patience for it right now, will have to do it later. I can just say that I have to understand the whole concept of abstract classes much better, I dont understand how I would be able to set specific member for the derived classes, I cant use there functions without downcasting ? I am quite confused with this and also with the rules when in inheritance a method shall be protected or private or public. Glad that the map together with the painted map seems to work though :).