Plan
I have envisioned, the playerlists shall be divided right and left from the board and the name of the player shall be the headline and highlighted in the respective color (or something else shows which color belongs to whom, maybe the dots of the list or the whole list?). I would also really like to have some images of dices at least to display when the dice is being rolled, maybe with a timer inbetween each dice to make it a bit more exciting. The country infos shall be displayed in a table which will be updated accordingly and which shall be opened via the menu. Further missing are the starting menu, maybe I can merge the choice of mission risk / normal risk and the player selection in one frame and dialog.