fix mission conquer 18 countries each with two armies
the issue was that the map in the mission class was working on different objects than the map which was updated in the business layer. Therefore, after reading from the persistence I added some lines in the initialize method in the playermanager in which the countryMap in the mission with number 4 is being set. I wonder if this is the best way, would be nicer to completely manage the mission creation in the persistence / in one method. I had the idea to introduce a class variable for the countrymap in the persistence, therefore the fetchCountries() method would only have to be called once (and it does not change, so no harm in that) and all the other persistence methods would work with this one map. But then, is it okay to have a class variable in the persistence? and I think for all other methods it does not matter, just for this one.

Apart from that the mission (after fixing the issues) returned true when it was supposed to and false when it was supposed to.