remove persistence from presentation layer
seems to be better to not have any import from the persistence into the presentation package. I do not know whether the getPersistence() method in worldManager is "allowed" but on the other side, why would it not be? Not sure if this is a good way or not but I think I will leave it like this for now.