introduction of SOUTH component label which displays text instead of messageboxes
my brother's suggestion, it's much better now I think.

Weirdly there is the issue that sometimes (but just sometimes) the playerPanel with the list goes blank when updated. Why? Then, when I test again, in the same situation it works. There must be an error somewhere? Also the dices are sometimes not being displayed. But it does not depend on the time when the dice is rolled or the dice number. These are problems more difficult to solve but there must be something wrong in the code, both regarding the playerPanel (or the countryList which is being changed?) and the method for displaying the dice images.