Update GameSetUpView.java
I'd really like to buy this book, just had a short glimpse in it in a bookstore. Two things stuck with me - 5 lines per method and always start with if (if it exists). Now, I would really like to know what's written in there about class attributes for example because now I created a lot. Should I have rather passed them in the parameter? Also I need to change the actionListener I think. It seems really impossible to have only five lines per method, maybe I even misunderstood because I didn't read the book that much. Would it not lead to way too many methods which would interrupt the flow of reading the code? Not sure if this here will lead to anything but I was just curious to try the 5 lines per method idea.